Split widget: devtools scope
