+++ extension/devtools/app/lib/devtools_scope.dart
@@ -1,0 +1,16 @@
+part of 'main.dart';
+
+class OrefDevToolsScope extends InheritedNotifier<OrefDevToolsController> {
+  const OrefDevToolsScope({
+    super.key,
+    required OrefDevToolsController controller,
+    required super.child,
+  }) : super(notifier: controller);
+
+  static OrefDevToolsController of(BuildContext context) {
+    final scope = context
+        .dependOnInheritedWidgetOfExactType<OrefDevToolsScope>();
+    assert(scope != null, 'OrefDevToolsScope not found in widget tree.');
+    return scope!.notifier!;
+  }
+}
